requirements.txt
flask
asyncio
vt-py
python-dotenv